third_party/blink/web_tests/transforms/2d/compound-2d-transforms-expected.txt

transform "none" expected "none" : PASS
transform "" expected "matrix(5.96046e-08, 1, -1, 5.96046e-08, 0, 0)" : PASS
transform "translate(50px, 20px)" expected "matrix(1, 0, 0, 1, 50, 20)" : PASS
transform "rotate(12deg)" expected "matrix(0.978148, 0.207912, -0.207912, 0.978148, 0, 0)" : PASS
transform "scale(1.7)" expected "matrix(1.7, 0, 0, 1.7, 0, 0)" : PASS
transform "skew(12deg)" expected "matrix(1, 0, 0.212557, 1, 0, 0)" : PASS
transform "matrix(0.978148, 0.207912, -0.207912, 0.978148, 50, 20)" expected "matrix(0.978148, 0.207912, -0.207912, 0.978148, 50, 20)" : PASS
transform "translate(50px, 20px) rotate(12deg)" expected "matrix(0.978148, 0.207912, -0.207912, 0.978148, 50, 20)" : PASS
transform "rotate(12deg) translate(50px, 20px)" expected "matrix(0.978148, 0.207912, -0.207912, 0.978148, 44.7491, 29.9585)" : PASS
transform "rotate(12deg) translate(50px, 20px) scale(1.4)" expected "matrix(1.36941, 0.291076, -0.291076, 1.36941, 44.7491, 29.9585)" : PASS
transform "scale(1.4) rotate(12deg) translate(50px, 20px) " expected "matrix(1.36941, 0.291076, -0.291076, 1.36941, 62.6488, 41.942)" : PASS
transform "scale(1.4) translate(50px, 20px) rotate(12deg)" expected "matrix(1.36941, 0.291076, -0.291076, 1.36941, 70, 28)" : PASS
transform "translate(50px, 20px) rotate(12deg) translateY(50px) scale(1.4)" expected "matrix(1.36941, 0.291076, -0.291076, 1.36941, 39.6044, 68.9074)" : PASS
transform "rotate(12deg) translate(50px, 20px) scale(1.4) translate(-80px, 40px) skew(34deg) translate(10px, 10px) scale(0.7) skewY(-25deg) rotate(21deg) translateX(50px) scale(1.4, 1.8)" expected "matrix(1.20517, 0.148207, 0.258987, 2.04004, -13.3847, 85.31)" : PASS
transform ",rotate(12deg) translate(50px, 20px)" expected "none" : PASS
transform "rotate(12deg), translate(50px, 20px)" expected "none" : PASS
transform "rotate(12deg) + translate(50px, 20px)" expected "none" : PASS
transform "rotate(12deg) translate(50, 20)" expected "none" : PASS
transform "(rotate(12deg) translate(50px, 20px))" expected "none" : PASS
transform "rotate(12deg) translate(50px, 20px) eggs(10)" expected "none" : PASS
transform "rotate(12deg) matrix(0.978148, 0.207912, -0.207912, 0.978148, 50, 20, 666)" expected "none" : PASS
